+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 
   * Added optional selector to findByTextContaining
   * Back, forward and refresh methods
+  * Settings updated if changed between browser creation
   * Support for node 8.0.0
   * Major code refactor
-  * Mixwith dependency added
+  * Mixwith, compositer and deep-equal dependencies added
 
 0.4.6 / 2018-03-24
 ==================